Fixed all tests.

+++ src/dcegm/pre_processing/setup_model.py
@@ -59,7 +59,7 @@ def setup_model(
     )
 
     model_funcs["exog_state_mapping"] = create_exog_state_mapping(
-        model_structure["exog_state_space"].astype(np.int16),
+        model_structure["exog_state_space"].astype(np.int64),
         model_structure["exog_states_names"],
     )
 
@@ -133,7 +133,7 @@ def load_and_setup_model(
 
     exog_state_space = model["model_structure"]["exog_state_space"]
     model["model_funcs"]["exog_state_mapping"] = create_exog_state_mapping(
-        exog_state_space=np.array(exog_state_space, dtype=np.int16),
+        exog_state_space=np.array(exog_state_space, dtype=np.int64),
         exog_names=model["model_structure"]["exog_states_names"],
     )
